Just tired fr

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -6,7 +6,7 @@
  *
  * Return: None.
  */
-void print_number(int n)
+void print_number(long n)
 {
 	if (n < 0)
 	{
